+++ src/main/java/es/in2/issuer/backend/shared/domain/service/impl/VerifiableCredentialServiceImpl.java
@@ -88,8 +88,7 @@ public class VerifiableCredentialServiceImpl implements VerifiableCredentialServ
             String authServerNonce,
             String token,
             String email) {
-        log.info("buildCredentialResponse");
-        log.info("email: {}", email);
+        log.debug("buildCredentialResponse - email: {} - processId: {}", email, processId);
         return deferredCredentialMetadataService
                 .getProcedureIdByAuthServerNonce(authServerNonce)
                 .flatMap(procedureId -> credentialProcedureService
@@ -148,8 +147,6 @@ public class VerifiableCredentialServiceImpl implements VerifiableCredentialServ
             String authServerNonce,
             String token,
             String email) {
-        log.info("updateDeferredAndMap");
-        log.info("email: {}", email);
 
         return deferredCredentialMetadataService
                 .updateDeferredCredentialMetadataByAuthServerNonce(authServerNonce)